Finch reference frame; setMove and setTurn waits; misc.
Finch now returns magnetometer and accelerometer data in it's own reference frame. Also added a wait to the resetEncoders function to give the finch a chance to reset before things move on. Finch setMove and setTurn functions now wait until the motion is complete before returning. Added a few options to what is acceptable input.